requirements.txt
anyio==3.3.4
asgiref==3.4.1
astroid==2.8.4
attrs==21.2.0
boto3==1.19.12
botocore==1.22.12
click==8.0.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
isort==5.9.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pydantic==1.8.2
pylint==2.11.1
pyparsing==3.0.3
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.11.1
requests-toolbelt==0.7.0
s3transfer==0.5.0
six==1.12.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wrapt==1.13.3
